refactor Node
